+++ src/main/java/com/blockguard/server/domain/analysis/domain/enums/RiskLevel.java
@@ -1,0 +1,5 @@
+package com.blockguard.server.domain.analysis.domain.enums;
+
+public enum RiskLevel {
+    Dangers, Caution, Safety
+}
