--- src/main/java/com/blockguard/server/domain/fraudanalysis/domain/enums/RiskLevel.java
@@ -1,5 +1,0 @@
-package com.blockguard.server.domain.fraudanalysis.domain.enums;
-
-public enum RiskLevel {
-    Dangers, Caution, Safety
-}
