+++ src/main/java/com/blockguard/server/domain/user/domain/User.java
@@ -1,6 +1,6 @@
 package com.blockguard.server.domain.user.domain;
 
-import com.blockguard.server.domain.fraudanalysis.domain.FraudAnalysisRecord;
+import com.blockguard.server.domain.analysis.domain.FraudAnalysisRecord;
 import com.blockguard.server.domain.guardian.domain.Guardian;
 import com.blockguard.server.domain.user.domain.enums.Gender;
 import com.blockguard.server.domain.report.domain.UserReportRecord;
